Add AddonsTest (install example addon). Add \gp\admin\Addon\Available
